requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.56.1
python-dotenv==1.0.1